requirements.txt
mujoco_py
gym==0.18.0
matplotlib==3.5.1
numpy==1.22.3
PyYAML==6.0
scikit_learn==1.1.2
scipy==1.8.0
tensorboardX==2.5.1
torch==1.11.0
tqdm==4.64.0
